+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - "inputs" field to substra.sdk.schemas.*tupleSpec (#215)
+- "inputs" field to *tuple models (#239)
 
 ## [0.30.1](https://github.com/owkin/substra/releases/tag/0.30.1) - 2022-07-26